requirements.txt
agent-framework
mcp
python-dotenv
jupyter
ipykernel
